Create pisound_ip.py
This python3 program runs periodically on the pi and finds the ip address and checks if a Microview is connected on USB and then sends the IP to the microview (to display).
It checks both eth0 and wlan0. If they are both connected and have ip address it will use the last one checked.